ArchUnit clean up and moved a class only used for test out of production.
