+++ geode-core/src/test/java/org/apache/geode/distributed/internal/membership/gms/api/MembershipAPIArchUnitTest.java
@@ -29,7 +29,7 @@ import org.junit.runner.RunWith;
 
 import org.apache.geode.distributed.DistributedMember;
 import org.apache.geode.distributed.internal.membership.InternalDistributedMember;
-import org.apache.geode.distributed.internal.membership.MembershipView;
+import org.apache.geode.distributed.internal.ClusterDistributionManager;
 import org.apache.geode.distributed.internal.membership.gms.MemberDataBuilderImpl;
 import org.apache.geode.distributed.internal.membership.gms.MembershipBuilderImpl;
 import org.apache.geode.distributed.internal.membership.gms.Services;
@@ -55,15 +55,12 @@ public class MembershipAPIArchUnitTest {
               // TODO: replace this with a rule allowing dependencies on geode-tcp-server module
               .or(type(TcpClient.class))
 
-              // this is allowed
+              // allowed
               .or(type(MembershipBuilderImpl.class))
               .or(type(MemberDataBuilderImpl.class))
 
               // TODO to be extracted as Interfaces
-              .or(type(InternalDistributedMember.class))
-              .or(type(MembershipView.class))
-              .or(type(DistributedMember.class))
-              .or(type(InternalDistributedMember[].class))
+              .or(type(ClusterDistributionManager.class))
 
               // TODO: This is used by the GMSLocatorAdapter to reach into the locator
               // part of the services
